`arithmetic` -> `algebra`
