+++ src/probnum/_function/__init__.py
@@ -1,2 +1,4 @@
-from ._arithmetic import ScaledFunction, SumFunction
+from ._algebra_fallbacks import ScaledFunction, SumFunction
 from ._function import Function, LambdaFunction
+
+from . import _algebra
